--- src/homeUser.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home User</title>
-    <link rel="stylesheet" type="text/css" href="./views/header.css">
-    <link rel="stylesheet" type="text/css" href="./views/main.css">
-    <link rel="stylesheet" type="text/css" href="./views/footer.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-
-
-
-</head>
-<body>
-    <div class="root" id="root"></div>
-    <script type="module"  src="./views/homeUser.js"></script>
-
-</body>
-</html>
