+++ src/index.html
@@ -3,17 +3,20 @@
 <head>
     <link href="<link rel=" preconnect " href=" https://fonts.googleapis.com ">
     <link rel=" preconnect " href=" https://fonts.gstatic.com " crossorigin>
-    <link href=" https://fonts.googleapis.com/css2?family=Playfair+Display&display=swap " rel=" stylesheet ">" <meta
-        charset="UTF-8">
+    <link href=" https://fonts.googleapis.com/css2?family=Playfair+Display&display=swap " rel=" stylesheet "> 
+    <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="views-css/style.css" />
+    <link rel="stylesheet" href="views-css/header.css" />
+    <link rel="stylesheet" href="views-css/mainPage.css" />
+    <link rel="stylesheet" href="views-css/footer.css" />
     <title>Terranova</title>
 </head>
 <body>
-    <div class="container" id="appContainer"></div>
-    <footer>
-    </footer>
+
+    <div id="appContainer"></div>
+
     <script type="module" src="main.js"></script>
 </body>
 </html>
