+++ src/lib/router.js
@@ -1,16 +1,16 @@
-import welcome from './views/welcome.js';
-import signin from './views/signin.js';
-import login from './views/login.js';
-import home from './views/home.js';
-import playing from './views/playing.js'
-import { navigateTo } from './lib/navigator.js';
+import welcome from '../views/welcome.js';
+import signin from '../views/signin.js';
+import login from '../views/login.js';
+import home from '../views/home.js';
+import homeUser from '../views/homeUser.js';
+import { navigateTo } from './navigator.js';
 
 const routes = {
   '/': welcome,
   '/signin': signin,
   '/login': login,
   '/home': home,
-  '/playing': playing,
+  '/homeUser': homeUser,
 };
 
 let appContainer;
@@ -23,13 +23,22 @@ export const displayView = (route) => {
     return;
   }
 
-  if (view.renderElement) {
+  // Actualizar appContainer segun la funcion render implementada en la vista
+  if (view.renderElements) {
+    appContainer.innerHTML = '';
+
+    const elements = view.renderElements();
+    elements.forEach((el) => {
+      appContainer.appendChild(el);
+    });
+  } else if (view.renderElement) {
     const el = view.renderElement();
     appContainer.innerHTML = '';
     appContainer.appendChild(el);
   } else {
     appContainer.innerHTML = view.render();
   }
+
   if (view.afterRender) {
     view.afterRender();
   }
