+++ src/views-css/style.css
@@ -2,10 +2,10 @@ body {
     font-family: 'Playfair Display', serif;
     padding: 0;
     margin: 0; 
-    display: flex;
+    /* display: flex;
     flex-wrap: wrap;
     justify-content: center;
-    width: 100%;
+    width: 100%; */
 
     background-image: url(https://www.ocu.org/-/media/ocu/images/home/vivienda%20y%20energia/energia%20renovable/hub_headerimage_v2.png?rev=70b9963a-74a8-4b0b-b93b-4f25ed81ed98&hash=2AC058B9B4A185A599B87F51658D18FC);
    
@@ -19,18 +19,25 @@ body {
     /* background-image: url(); */
     /* background-image: url(); */
 }      
-.container{
+/* .container{
     width:60%;
-} 
+}  */
 
 .title {
     text-align: center;
     color: rgb(14, 31, 51);
     font-size: 40px;
 }
+
  .img-fluid {
     display: flex;
      justify-content: center;
      margin: 0 auto;
  }
+
+ .logInSigninEmail {
+    margin: auto;
+    width: 30rem;
+    max-width: 80%;
+   }
